requirements.txt
python-dotenv==1.1.1

fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0

cryptography==45.0.6
tortoise-orm[aiomysql]==0.25.1